requirements-core.txt
fastapi~=0.42.0
uvicorn~=0.9.0 
etcd3~=0.10.0
starlette-prometheus~=0.5.0
func-timeout~=4.3.5
